Loop over picasa files instead of directories for robustness
If we loop over dirs and find picasa files, we might get multiple or no picasa files, resulting in
errors. If we find all picasa files and loop over those, we're alsways sure we get exactly one. For
dirs without picasa file we won't find anything and skip it (OK), for dirs with two or more files
we'll parse them sequentially (weird but also OK).